add setup.py targets to build sphinx documentation
It is now possible to run: `python setup.py build_sphinx`
which takes care of installing all requirements.

`python setup.py test` command must be ran before so that
shared library is properly installed in basalt/ directory.

This patch also adds the standard `upload_docs` target to
upload the documentation to https://readthedocs.org/
but basalt is not configured on rtd so it is currently failing.